Add locationname automatically to form
